requirements.txt
requests
readability-lxml
beautifulsoup4
Pillow
mutagen
pyperclip
ddgs